requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings>=2.1.0
python-multipart==0.0.6

# ML — optional; service degrades gracefully without these
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
scikit-learn>=1.4.0
httpx>=0.27.0

# NLP/LLM integration
anthropic>=0.25.0
python-dotenv>=1.0.0

# ML extras (install manually or via Docker for full FOHEM support)
# scikit-fuzzy>=0.4.2
# catboost>=1.2.3
# deap>=1.4.1
# shap>=0.44.0
# lime>=0.2.0.1
# influxdb-client>=1.40.0
# mlflow>=2.11.0
# alembic>=1.13.0
# celery>=5.3.6
# redis>=5.0.3
# geopandas>=0.14.0
# rasterio>=1.3.9
# psycopg2-binary>=2.9.9
